awk the docker host ip address
Previously the NGINX_HOST environment variable was just set to the default from the book. Not sure why the reverse proxy worked with this host ip. The 45.55.250.0/24 CIDR is owned by Digital Ocean but I believe the relationship is orthogonal. Next thing to do would be enabling logging for nginx and set severity to warning.

Security or performance concerns with logging in production?  Does it still work with an invalid proxy IP?

Signed-off-by: Jacob Chaffin <jchaffin57@gmail.com>